pila: add method to present stacks on a key-value mode
