+++ pila/stack_test.go
@@ -2,7 +2,6 @@ package pila
 
 import (
 	"bytes"
-	"fmt"
 	"reflect"
 	"testing"
 )
@@ -159,169 +158,6 @@ func TestStackSizeToJSON(t *testing.T) {
 	expectedSize := `3`
 	if string(stack.SizeToJSON()) != expectedSize {
 		t.Errorf("size is %s, expected %s", string(stack.SizeToJSON()), expectedSize)
-	}
-}
-
-func TestStackStatusJSON(t *testing.T) {
-	stack := NewStack("test-stack")
-	stack.Push("test")
-	stack.Push(8)
-	stack.Push(5.87)
-	stack.Push([]byte("test"))
-
-	expectedStatus := fmt.Sprintf(`{"id":"2f44edeaa249ba81db20e9ddf000ba65","name":"test-stack","peek":"dGVzdA==","size":4}`)
-	if status, err := stack.Status().ToJSON(); err != nil {
-		t.Fatal(err)
-	} else if string(status) != expectedStatus {
-		t.Errorf("status is %s, expected %s", string(status), expectedStatus)
-	}
-}
-
-func TestStackStatusJSON_Empty(t *testing.T) {
-	stack := NewStack("test-stack")
-
-	expectedStatus := fmt.Sprintf(`{"id":"2f44edeaa249ba81db20e9ddf000ba65","name":"test-stack","peek":null,"size":0}`)
-	if status, err := stack.Status().ToJSON(); err != nil {
-		t.Fatal(err)
-	} else if string(status) != expectedStatus {
-		t.Errorf("status is %s, expected %s", string(status), expectedStatus)
-	}
-}
-
-func TestStackStatusJSON_Error(t *testing.T) {
-	// From https://golang.org/src/encoding/json/encode.go?s=5438:5481#L125
-	// Channel, complex, and function values cannot be encoded in JSON.
-	// Attempting to encode such a value causes Marshal to return
-	// an UnsupportedTypeError.
-
-	ch := make(chan int)
-	f := func() string { return "a" }
-
-	stack := NewStack("test-stack-channel")
-	stack.Push(ch)
-
-	if _, err := stack.Status().ToJSON(); err == nil {
-		t.Error("err is nil, expected UnsupportedTypeError")
-	}
-
-	stack = NewStack("test-stack-function")
-	stack.Push(f)
-
-	if _, err := stack.Status().ToJSON(); err == nil {
-		t.Error("err is nil, expected UnsupportedTypeError")
-	}
-}
-
-func TestStacksStatusJSON(t *testing.T) {
-	stack1 := NewStack("test-stack-1")
-	stack1.Push("test")
-	stack1.Push(8)
-	stack1.Push(5.87)
-	stack1.Push([]byte("test"))
-
-	stack2 := NewStack("test-stack-2")
-	stack2.Push("foo")
-	stack2.Push([]byte("bar"))
-	stack2.Push(999)
-
-	stacksStatus := StacksStatus{
-		Stacks: []StackStatus{stack1.Status(), stack2.Status()},
-	}
-
-	expectedStatus := fmt.Sprintf(`{"stacks":[{"id":"a0bfff209889f6f782997a7bd5b3d536","name":"test-stack-1","peek":"dGVzdA==","size":4},{"id":"f0d682fdfb3396c6f21e6f4d1d0da1cd","name":"test-stack-2","peek":999,"size":3}]}`)
-	if status, err := stacksStatus.ToJSON(); err != nil {
-		t.Fatal(err)
-	} else if string(status) != expectedStatus {
-		t.Errorf("status is %s, expected %s", string(status), expectedStatus)
-	}
-}
-
-func TestStacksStatusJSON_Empty(t *testing.T) {
-	stacksStatus := StacksStatus{
-		Stacks: []StackStatus{},
-	}
-
-	expectedStatus := fmt.Sprintf(`{"stacks":[]}`)
-	if status, err := stacksStatus.ToJSON(); err != nil {
-		t.Fatal(err)
-	} else if string(status) != expectedStatus {
-		t.Errorf("status is %s, expected %s", string(status), expectedStatus)
-	}
-}
-
-func TestStacksStatusJSON_Error(t *testing.T) {
-	// From https://golang.org/src/encoding/json/encode.go?s=5438:5481#L125
-	// Channel, complex, and function values cannot be encoded in JSON.
-	// Attempting to encode such a value causes Marshal to return
-	// an UnsupportedTypeError.
-
-	ch := make(chan int)
-	f := func() string { return "a" }
-
-	stack := NewStack("test-stack-channel")
-	stack.Push(ch)
-
-	stacksStatus := StacksStatus{
-		Stacks: []StackStatus{stack.Status()},
-	}
-
-	if _, err := stacksStatus.ToJSON(); err == nil {
-		t.Error("err is nil, expected UnsupportedTypeError")
-	}
-
-	stack = NewStack("test-stack-function")
-	stack.Push(f)
-
-	stacksStatus = StacksStatus{
-		Stacks: []StackStatus{stack.Status()},
-	}
-
-	if _, err := stacksStatus.ToJSON(); err == nil {
-		t.Error("err is nil, expected UnsupportedTypeError")
-	}
-}
-
-func TestStacksStatusLen(t *testing.T) {
-	stack1 := NewStack("test-stack-1")
-	stack2 := NewStack("test-stack-2")
-
-	stacksStatus := StacksStatus{
-		Stacks: []StackStatus{stack1.Status(), stack2.Status()},
-	}
-
-	expectedLen := 2
-	if len := stacksStatus.Len(); len != expectedLen {
-		t.Errorf("len is %d, expected %d", len, expectedLen)
-	}
-}
-
-func TestStacksStatusLess(t *testing.T) {
-	stack1 := NewStack("test-stack-1")
-	stack2 := NewStack("test-stack-2")
-
-	stacksStatus := StacksStatus{
-		Stacks: []StackStatus{stack1.Status(), stack2.Status()},
-	}
-
-	if less := stacksStatus.Less(0, 1); !less {
-		t.Errorf("less is %v, expected %v", less, true)
-	}
-}
-
-func TestStacksStatusSwap(t *testing.T) {
-	stack1 := NewStack("test-stack-1")
-	stack2 := NewStack("test-stack-2")
-
-	stacksStatus := StacksStatus{
-		Stacks: []StackStatus{stack1.Status(), stack2.Status()},
-	}
-
-	expectedStacksStatus := StacksStatus{
-		Stacks: []StackStatus{stack2.Status(), stack1.Status()},
-	}
-
-	if stacksStatus.Swap(0, 1); !reflect.DeepEqual(stacksStatus, expectedStacksStatus) {
-		t.Errorf("status is %v, expected %v", stacksStatus, expectedStacksStatus)
 	}
 }
